Put spaces in blankqueue too, popping if we write a word with spaces
and only print *one* item from blankqueue in printSpace.

This fixes #47. With the input 'legge opp<br/>blah', we'll now have
the regular space in blankqueue and use that before getting to the
<br/>.

Tested on 100.000 lines of html with nno-nob, only good changes found,
and seems to work fine with flushing.