Replace some `std::set<int>` with a sorted `std::vector`
`Transducer::determinize()` doesn't do very much inserting into sets,
but does iterate over them and use them as map keys, both operations
which are substantially sped up by ensuring that the elements of the
set are contiguous. Thus this commit adds `sorted_vector` which is a wrapper
around `std::vector` that ensures that inserted elements are sorted
and unique. The results are significant improvements in both runtime
and memory usage for transducer minimization.

Effect on `lt-comp`:

|          | eng           | oci           |
|----------|---------------|---------------|
| orig     | 136 MB 23.9 s |  980 MB 219 s |
| pre-comp | 160 MB 16.8 s | 1120 MB 163 s |
| VecSet   |  99 MB 12.7 s |  800 MB 116 s |
| net diff | -27%   -47%   | -18%    -47%  |