Avoid string copies and comparisons in `Transducer::intersect()`
Determining whether transitions in different FSTs are the same symbol
can be done without converting to strings, special symbols can be
pre-computed, and string comparisons for tags can be done in-place by
the alphabets to eliminate the copying. This leads to a roughly 10%
speedup when running `lt-trim` on eng-cat.

Also replace `std::list` with `std::vector`.