add new tool lt-compose to compose fst binaries
By default `lt-compose f g gf` applies the input-side (left) of `g` to
the output-side (right) of `f`. But you can use `--inverted` to apply
`g` to the *input*-side of `f`. You would do this when altering the
forms of an analyser.

By default, we require matching full paths from initial to final. But
you can use `--anywhere` to allow the second fst to apply wherever in
the paths of the first (as if it were surrounded by ".*" on both
sides). Composition is then optional – the output fst is a superset of
the first one.

Runs on the sections of `f` in parallel if LT_JOBS=yes.

When arcs compose, their weights are added.

The alphabet of `f` is re-used and expanded/altered where needed with
symbols from `g`.

Uses ax_pthread.m4 to check for phtread support, from
http://git.savannah.gnu.org/gitweb/?p=autoconf-archive.git;a=blob_plain;f=m4/ax_pthread.m4