$ echo '^ikke/ikke<adv>$ ^«/«<lquot><MERGE_BEG>$^så/så<adv>$ ^ve\[dig/v<adv>$^»/»<rquot><MERGE_END>$ ^bra/bra<adj>$' | lttoolbox/lt-proc  -b ../apertium-nno-nob/nob-nno.autobil.bin 2>/dev/null
^ikke/ikke<adv>$ ^«så ve\[dig»/«så ve\[dig»<MERGED>$ ^bra/bra<adj>$

$ echo '^«/«<lquot><MERGE_BEG>$[[tf:i:a]]^ve\/ldig/v<adv>$[[/]]^»/»<rquot><MERGE_END>$' | lttoolbox/lt-proc  -b ../apertium-nno-nob/nob-nno.autobil.bin 2>/dev/null
^«\[\[tf:i:a\]\]ve\/ldig\[\[\/\]\]»/«\[\[tf:i:a\]\]ve\/ldig\[\[\/\]\]»<MERGED>$

TODO:

- put it into a command of its own lt-merge

- need to be able to pass MERGED stuff unchanged through biltrans and
  generator, would like to <re>.+</re><i><s n="MERGED"/></i> but . is
  literal period in re(!) and even ANY_CHAR doesn't seem supported in
  lt-proc -b

QUESTIONS:

If we run this between analysis and wblank-attach, then after
the `lt-proc -b generator.bin` step we have e.g.

    ^ikke<adv>/ikke$ ^«\[\[tf:i:a\]\]så\[\[\/\]\]»<MERGED>/«\[\[tf:i:a\]\]så\[\[\/\]\]»$

which after `cg-proc -1 -n -g genprefs.bin` would turn into

    ikke «\[\[tf:i:a\]\]så\[\[\/\]\]»

We want that wblank unquoted for tf-injection here, but it seems hacky
to just match all `\[\[tf` in the plain stream. Perhaps we run a
command `lt-unmerge` which looks for `<MERGED>` tags and turns the
generator output (still in apertium stream format) into

    ^ikke<adv>/ikke$ «[[tf:i:a]]så[[/]]»

This will go fine through cg-proc -g, but the initial `lt-merge` would
have to triple-quote anything that's *not* word-blank. E.g. a ^ should
become \\\^ so that

    $ echo '^ikke<adv>/ikke$ ^«\[\[tf:i:a\]\]s\\\^å\[\[\/\]\]»<MERGED>/«\[\[tf:i:a\]\]s\\\^å\[\[\/\]\]»$' | lt-unmerge-TODO
    ^ikke<adv>/ikke$ «[[tf:i:a]]s\^å[[/]]»

which then becomes

    $ echo '^ikke<adv>/ikke$ «[[tf:i:a]]s\^å[[/]]»' |cg-proc -1ng nob-nno.genprefs.rlx.bin
    ikke «[[tf:i:a]]s\^å[[/]]»

which tf-inject is happy to handle.